move LLM client related code to llm.go
